Fix gazebo crash on plugin destructor
Gazebo was crashing when the model associated with the plugin was deleted. The
problem was solved after removing the DisconnectWorldUpdateBegin() call on the
plugin destructor. This call is never used in any gazebo plugin and has been
made deprecated on Gazebo 8, what makes me think that it's a common source of
problems. The regular destructor for the update event should be used instead.

Signed-off-by: Guilherme Campos Camargo <guilherme.campos.camargo@intel.com>