Remove Subs target
Subs target was introduced to the plugin to show in the interface what
was the detour waypoint set by the coav system. This was never
implemented on the coav side though. It also assumes that the plugin
will in the middle between coav and the vehicle, but in our current
architecture coav communicates to the vehicle directly.