Wait for takeoff
If we change the vehicle mode back to AUTO right after a takeoff
command it will move right away and may crash or be unable to identify
obstacles due to the proximity to the ground.

This patch waits until the vehicle is at least 1m high before before
proceeding to AUTO mode.

Signed-off-by: Rodrigo Chiossi <rodrigo.chiossi@intel.com>